error handling on supplier and sales panels

+++ src/pos_java_ant_1/Supplier.java
@@ -9,7 +9,9 @@ import java.io.InputStream;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.sql.ResultSet;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Vector;
 import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
@@ -26,6 +28,7 @@ public class Supplier extends javax.swing.JPanel {
     public Supplier() {
         initComponents();
         tableLoad();
+        
     }
     
     public void tableLoad(){
@@ -597,8 +600,8 @@ public class Supplier extends javax.swing.JPanel {
             jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel4Layout.createSequentialGroup()
                 .addContainerGap()
-                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 992, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane1)
                     .addGroup(jPanel4Layout.createSequentialGroup()
                         .addComponent(jLabel5)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
@@ -614,8 +617,9 @@ public class Supplier extends javax.swing.JPanel {
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                         .addComponent(jLabel17)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(csearchtblemail)))
-                .addContainerGap(133, Short.MAX_VALUE))
+                        .addComponent(csearchtblemail, javax.swing.GroupLayout.PREFERRED_SIZE, 138, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 123, Short.MAX_VALUE)))
+                .addContainerGap())
         );
         jPanel4Layout.setVerticalGroup(
             jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -722,7 +726,7 @@ public class Supplier extends javax.swing.JPanel {
         this.setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jTabbedPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 1140, Short.MAX_VALUE)
+            .addComponent(jTabbedPane1)
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -735,7 +739,9 @@ public class Supplier extends javax.swing.JPanel {
 
     private void jButton6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton6ActionPerformed
         // TODO add your handling code here:
-        try {
+        String search = report_cid.getText();
+        if (!search.isEmpty() && search.matches("[0-9]{1,4}")) {
+            try {
             HashMap map = new HashMap();
             map.put("para_sid", report_cid.getText());
 
@@ -744,6 +750,12 @@ public class Supplier extends javax.swing.JPanel {
         } catch (Exception e) {
             JOptionPane.showMessageDialog(null, "Somethings wrong in your system!");
         }
+        } else {
+            JOptionPane.showMessageDialog(null, "Enter valid ID");
+            report_cid.selectAll();
+            report_cid.requestFocus();
+        }
+        
         
     }//GEN-LAST:event_jButton6ActionPerformed
 
@@ -921,14 +933,42 @@ public class Supplier extends javax.swing.JPanel {
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
         // TODO add your handling code here:
-        save();
+        List<String> formFields = new ArrayList<>();
+        formFields.add(cname.getText());
+        formFields.add(ctp.getText());
+        formFields.add(addr.getText());
+        formFields.add(type.getText());
+        formFields.add(bank.getText());
+        formFields.add(city.getText());
+        formFields.add(cname2.getText());
+        formFields.add(con_name.getText());
+        formFields.add(con_tp.getText());
+        formFields.add(con_email.getText());
+        formFields.add(con_online.getText());
+        
+        boolean allFieldsFilled = true;
+
+        for (String field : formFields) {
+            if (field.isEmpty() || field.equals("0")) {
+                allFieldsFilled = false;
+                break;
+            }
+        }
+        
+        if (allFieldsFilled) {
+            save();
+        } else {
+            JOptionPane.showMessageDialog(null, "Please fill all the details");
+        }
+        
     }//GEN-LAST:event_jButton1ActionPerformed
 
     private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
         // TODO add your handling code here:
         String search = csearch.getText();
 
-        try {
+        if (!search.isEmpty() && search.matches("[0-9]{1,4}")) {
+            try {
             Statement s = db.myCon().createStatement();
             ResultSet rs = s.executeQuery(" SELECT * FROM supplier WHERE sid = '"+search+"'");
 
@@ -967,6 +1007,12 @@ public class Supplier extends javax.swing.JPanel {
         } catch (SQLException e) {
             System.out.println(e);
         }
+        } else {
+            JOptionPane.showMessageDialog(null, "Please enter valid user ID");
+            csearch.selectAll();
+            csearch.requestFocus();
+        }
+        
     }//GEN-LAST:event_jButton2ActionPerformed
 
     private void jButton4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton4ActionPerformed
@@ -983,8 +1029,31 @@ public class Supplier extends javax.swing.JPanel {
         String c_name = cname2.getText();
         String c_email = con_email.getText();
         String c_online = con_online.getText();
+        
+        List<String> formFields = new ArrayList<>();
+        formFields.add(cname.getText());
+        formFields.add(ctp.getText());
+        formFields.add(addr.getText());
+        formFields.add(type.getText());
+        formFields.add(bank.getText());
+        formFields.add(city.getText());
+        formFields.add(cname2.getText());
+        formFields.add(con_name.getText());
+        formFields.add(con_tp.getText());
+        formFields.add(con_email.getText());
+        formFields.add(con_online.getText());
+        
+        boolean allFieldsFilled = true;
 
-        try {
+        for (String field : formFields) {
+            if (field.isEmpty() || field.equals("0")) {
+                allFieldsFilled = false;
+                break;
+            }
+        }
+        
+        if (allFieldsFilled) {
+            try {
             Statement s = db.myCon().createStatement();
             s.executeUpdate(" UPDATE supplier SET name='"+name+"', tp='"+tp+"', addr='"+adr+"', shipping_addr='"+ship+"', bank='"+bnk+"', city='"+cty+"', c_name='"+c_name+"', c_per='"+c_per+"', c_tp='"+c_tp+"', c_email='"+c_email+"', c_online='"+c_online+"' WHERE sid='"+id+"'");
             cname.setText("");
@@ -1004,13 +1073,19 @@ public class Supplier extends javax.swing.JPanel {
         } catch (SQLException e) {
             System.out.println(e);
         }
+        } else {
+            JOptionPane.showMessageDialog(null, "Please fill all the details");
+        }
+
+        
     }//GEN-LAST:event_jButton4ActionPerformed
 
     private void jButton3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton3ActionPerformed
         // TODO add your handling code here:
         String id = csearch.getText();
 
-        try {
+        if (!id.isEmpty() && id.matches("[0-9]{1,4}")) {
+            try {
             Statement s = db.myCon().createStatement();
             s.executeUpdate(" DELETE FROM supplier WHERE sid='"+id+"'");
             cname.setText("");
@@ -1030,6 +1105,12 @@ public class Supplier extends javax.swing.JPanel {
         } catch (SQLException e) {
             System.out.println(e);
         }
+        } else {
+            JOptionPane.showMessageDialog(null, "Please enter valid ID");
+            csearch.selectAll();
+            csearch.requestFocus();
+        }
+        
     }//GEN-LAST:event_jButton3ActionPerformed
 
     private void con_emailKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_con_emailKeyReleased
